refactor control components into separate files
